fix(build): use known working node version
- fix npm links fallback
- fix npm links handling for installer too
- use matching npm version for node version when building
- GitHub workflows: allow triggering of installer creation without tagging
