+++ .github/workflows/release.yml
@@ -5,6 +5,8 @@ on:
     tags:
       - 'v[0-9]+.[0-9]+.[0-9]+'
       - 'v[0-9]+.[0-9]+.[0-9]+-*'
+  workflow_dispatch:
+
 
 jobs:
   run-build:
@@ -23,7 +25,14 @@ jobs:
         run: Copy-Item (Get-Command node.exe | Select-Object -ExpandProperty Definition) .
       - run: npm test
       - run: npm run build
+      - name: Archive build artifact
+        if: github.ref_type == 'branch'
+        uses: actions/upload-artifact@v3
+        with:
+          name: Installer
+          path: "build/out/NodistSetup-*.exe"
       - name: Create release draft
+        if: github.ref_type == 'tag'
         uses: ncipollo/release-action@v1
         with:
           artifacts: "build/out/NodistSetup-*.exe"
