+++ lib/build.js
@@ -4,7 +4,11 @@ var path = require('path');
 var promisePipe = require('promisepipe');
 var ProgressBar = require('progress');
 var request = require('request');
+const P = require('bluebird');
 var debug = require('debug')('nodist:build')
+
+//make some promising APIs
+P.promisifyAll(fs);
 
 /**
  * Copy File
@@ -141,3 +145,51 @@ exports.downloadFileStream = function downloadFileStream(url) {
   });
   return req
 }
+
+/**
+ * Npm version >= 18 using symlinks that do not work in windows and have to be fixed
+ * this function replace the broken symlinks with NTFS junction or move the directory if junctions are not supported
+ *
+ * @param {string} dirPath
+ * @param {boolean} preferSymlink
+ * @returns {Promise<number>} number of changed links
+ */
+exports.resolveLinkedWorkspaces = async function resolveLinkedWorkspaces(dirPath, preferSymlink = true) {
+  let fixedLinks = 0;
+  const packageLockJson = JSON.parse(fs.readFileSync(path.join(dirPath, 'package-lock.json')).toString());
+  await Promise.all(Object.entries(packageLockJson.packages)
+    .filter(([pkgPath, pkg]) => pkg.link === true)
+    .map(async ([pkgPath, pkg]) => {
+
+      const linkPath = path.join(dirPath, pkgPath);
+
+
+      if (await fs.accessAsync(linkPath, fs.constants.F_OK).then(() => true).catch(() => false)) {
+        await fs.unlinkAsync(linkPath);
+      }
+
+      let linkCreated = false;
+      if (preferSymlink) {
+        const linkTarget = path.join(
+          ...pkgPath.split('/').slice(0, -1).map(() => '..'),
+          pkg.resolved
+        );
+        debug('Create symlink for ', linkPath, 'with target', linkTarget);
+        try {
+          await fs.symlinkAsync(linkTarget, linkPath, 'junction');
+          linkCreated = true;
+        } catch (e) {
+          debug('Link ', linkPath, 'could not be created');
+        }
+      }
+      if (!linkCreated) {
+        const from = path.join(dirPath, pkg.resolved);
+        debug('Move', from, 'to', linkPath);
+        await fs.renameAsync(from, linkPath);
+      }
+
+      fixedLinks++;
+    }));
+
+  return fixedLinks;
+};
