+++ lib/npm.js
@@ -37,41 +37,6 @@ module.exports = npmist
 var NPMIST = npmist.prototype
 
 /**
- * Npm version >= 18 using symlinks that do not work in windows and have to be fixed
- * this function replace the broken symlinks with NTFS junction or move the directory if junctions are not supported
- *
- * @param {string} dirPath
- * @returns {Promise<number>} number of changed links
- */
-async function resolveLinkedWorkspaces(dirPath) {
-  let fixedLinks = 0;
-  const packageLockJson = JSON.parse(fs.readFileSync(path.join(dirPath, 'package-lock.json')).toString());
-  await Promise.all(Object.entries(packageLockJson.packages)
-    .filter(([pkgPath, pkg]) => pkg.link === true)
-    .map(async ([pkgPath, pkg]) => {
-      const linkTarget = path.join(
-        ...pkgPath.split('/').slice(0, -1).map(() => '..'),
-        pkg.resolved
-      );
-      const linkPath = path.join(dirPath, pkgPath);
-
-      debug('Create symlink for ', linkPath, 'with target', linkTarget);
-      if (await fs.accessAsync(linkPath, fs.constants.F_OK).then(() => true).catch(() => false)) {
-        await fs.unlinkAsync(linkPath);
-      }
-
-      try {
-        await fs.symlinkAsync(linkTarget, linkPath, 'junction');
-      } catch (e) {
-        await fs.renameAsync(path.join(dirPath, linkTarget), linkPath);
-      }
-      fixedLinks++;
-    }));
-
-  return fixedLinks;
-}
-
-/**
  * List available NPM versions
  * @return {string}
  */
@@ -331,7 +296,7 @@ NPMIST.install = function(v,done){
   .then(() => {
     if (semver.gte(version, '8.0.0')) {
       debug('Fix symlinks for npm version >= 8');
-      return resolveLinkedWorkspaces(path.join(archivePath))
+      return buildHelper.resolveLinkedWorkspaces(path.join(archivePath))
         .then(fixedLinks => {
           debug(`Fixed ${fixedLinks} symlinks for npm node_modules`);
         });
